// apps/frontend/app/api/v1/osograph/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
  [P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: string; output: string };
  JSON: { input: Record<string, unknown>; output: Record<string, unknown> };
};
// ...
export type DataConnection = {
  __typename?: "DataConnection";
  config: Scalars["JSON"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  type: DataConnectionType;
  updatedAt: Scalars["DateTime"]["output"];
};

export type DataConnectionAlias = {
  __typename?: "DataConnectionAlias";
  dataConnectionId: Scalars["ID"]["output"];
  datasetId: Scalars["ID"]["output"];
  id: Scalars["ID"]["output"];
  materializations: MaterializationConnection;
  modelContext?: Maybe<ModelContext>;
  orgId: Scalars["ID"]["output"];
  previewData: PreviewData;
  schema: Scalars["String"]["output"];
};

export type DataConnectionAliasMaterializationsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  tableName: Scalars["String"]["input"];
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DataConnectionAliasModelContextArgs = {
  tableName: Scalars["String"]["input"];
};

export type DataConnectionAliasPreviewDataArgs = {
  tableName: Scalars["String"]["input"];
};

export type DataConnectionConnection = {
  __typename?: "DataConnectionConnection";
  edges: Array<DataConnectionEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export const DataConnectionType = {
  Bigquery: "BIGQUERY",
  Gsheets: "GSHEETS",
  Postgresql: "POSTGRESQL",
} as const;

export type DataConnectionType =
  (typeof DataConnectionType)[keyof typeof DataConnectionType];
// ...
export type DatasetConnection = {
  __typename?: "DatasetConnection";
  edges: Array<DatasetEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type MaterializationConnection = {
  __typename?: "MaterializationConnection";
  edges: Array<MaterializationEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type MaterializationEdge = {
  __typename?: "MaterializationEdge";
  cursor: Scalars["String"]["output"];
  node: Materialization;
};

export const MemberRole = {
  Admin: "admin",
  Owner: "owner",
} as const;

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole];
export type ModelColumnContext = {
  __typename?: "ModelColumnContext";
  context?: Maybe<Scalars["String"]["output"]>;
  name: Scalars["String"]["output"];
};

export type ModelColumnContextInput = {
  context?: InputMaybe<Scalars["String"]["input"]>;
  name: Scalars["String"]["input"];
};

export type ModelContext = {
  __typename?: "ModelContext";
  columnContext?: Maybe<Array<ModelColumnContext>>;
  context?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  datasetId: Scalars["ID"]["output"];
  id: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
  tableId: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type NotebookConnection = {
  __typename?: "NotebookConnection";
  edges: Array<NotebookEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type NotebookEdge = {
  __typename?: "NotebookEdge";
  cursor: Scalars["String"]["output"];
  node: Notebook;
};

export type Organization = {
  __typename?: "Organization";
  createdAt: Scalars["DateTime"]["output"];
  dataConnections: DataConnectionConnection;
  datasets: DatasetConnection;
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  members: UserConnection;
  name: Scalars["String"]["output"];
  notebooks: NotebookConnection;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  endCursor?: Maybe<Scalars["String"]["output"]>;
  hasNextPage: Scalars["Boolean"]["output"];
  hasPreviousPage: Scalars["Boolean"]["output"];
  startCursor?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type PreviewData = {
  __typename?: "PreviewData";
  isAvailable: Scalars["Boolean"]["output"];
  rows: Array<Scalars["JSON"]["output"]>;
};
// ...
export type UserConnection = {
  __typename?: "UserConnection";
  edges: Array<UserEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<
  TResult,
  TParent = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
  TArgs = Record<PropertyKey, never>,
> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo,
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<
  T = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
> = (
  obj: T,
  context: TContext,
  info: GraphQLResolveInfo,
) => boolean | Promise<boolean>;
// ...
export type DataConnectionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataConnection"] = ResolversParentTypes["DataConnection"],
> = {
  config?: Resolver<ResolversTypes["JSON"], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  type?: Resolver<
    ResolversTypes["DataConnectionType"],
    ParentType,
    ContextType
  >;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type DataConnectionAliasResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataConnectionAlias"] = ResolversParentTypes["DataConnectionAlias"],
> = {
  dataConnectionId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  datasetId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  materializations?: Resolver<
    ResolversTypes["MaterializationConnection"],
    ParentType,
    ContextType,
    RequireFields<
      DataConnectionAliasMaterializationsArgs,
      "first" | "tableName"
    >
  >;
  modelContext?: Resolver<
    Maybe<ResolversTypes["ModelContext"]>,
    ParentType,
    ContextType,
    RequireFields<DataConnectionAliasModelContextArgs, "tableName">
  >;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  previewData?: Resolver<
    ResolversTypes["PreviewData"],
    ParentType,
    ContextType,
    RequireFields<DataConnectionAliasPreviewDataArgs, "tableName">
  >;
  schema?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
};